requirements.txt
aiogram==2.25.2
apscheduler
python-dotenv
dateparser
